IPhone Stress Test (#322)
Simulates two days (10 sessions) of heavy, back-to-back data collection on the iPhone to try and surface any errors we may see in production. Task times for each session are based on percentiles of historical data in log_task: every session is at least the 50th percentile and a two are the 99th percentile and the 99th percentile + 20% extra. The data dump is only performed after all recordings are complete.

Also includes some minor updates to the test script functionality in iphone.py.

Overall, the stress test will take at least 5h 34 m to run. (Strictly longer, as this does not include the dump or task-to-task overhead.) Therefore, this stress test is intended to be started on staging ACQ and left alone for a day/night.