Updates to config files and config file management (#401)
* initial work on task_params

Note: Includes fixes to the task param files that need to be deployed to the configs repo and then to all servers

* Moved environment-specific config vals to a separate file, added logic to include them in model

* Update passage_demo_obs_1.yml

remove dupe eyelink reference

* Move the code reading the environment specific vars

This makes it get read once per folder, instead of once per file.  This isn't optimal, but isn't bad.  The benefit of putting it here is that it is sufficiently low level that it can't easily be bypassed in future code changes.